+++ example/lib/example5_themes_playground/theme/flex_theme_light.dart
@@ -432,6 +432,7 @@ FlexColorScheme flexColorSchemeLight(ThemeController controller, Color source) {
             cardBackgroundSchemeColor: controller.cardBackgroundSchemeColor,
             cardBorderSchemeColor: controller.cardBorderSchemeColor,
             cardBorderWidth: controller.cardBorderWidth,
+            cardElevation: controller.cardElevation,
             // PopupMenuButton settings.
             popupMenuRadius: controller.popupMenuBorderRadius,
             popupMenuElevation: controller.popupMenuElevation,